+++ flink_java/pom.xml
@@ -11,8 +11,9 @@
     <properties>
         <flink.version>1.13.5</flink.version>
         <java.version>1.8</java.version>
-        <scala.binary.version>2.12</scala.binary.version>
+        <scala.binary.version>2.11</scala.binary.version>
         <slf4j.version>1.7.30</slf4j.version>
+        <jedis.version>2.8.0</jedis.version>
     </properties>
 
     <dependencies>
@@ -131,6 +132,23 @@
             <version>4.3.1</version>
         </dependency>
 
+        <!-- https://mvnrepository.com/artifact/org.apache.flink/flink-s3-fs-hadoop -->
+        <dependency>
+            <groupId>org.apache.flink</groupId>
+            <artifactId>flink-s3-fs-hadoop</artifactId>
+            <version>${flink.version}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- https://mvnrepository.com/artifact/org.apache.flink/flink-s3-fs-presto -->
+        <dependency>
+            <groupId>org.apache.flink</groupId>
+            <artifactId>flink-s3-fs-presto</artifactId>
+            <version>${flink.version}</version>
+            <scope>provided</scope>
+        </dependency>
+
+
 
         <dependency>
             <groupId>com.alibaba</groupId>
@@ -158,6 +176,33 @@
             <artifactId>kafka_2.12</artifactId>
             <version>2.8.0</version>
         </dependency>
+
+
+
+
+
+
+        <dependency>
+            <groupId>redis.clients</groupId>
+            <artifactId>jedis</artifactId>
+            <version>${jedis.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+            <version>${slf4j.version}</version>
+        </dependency>
+
+
+
+
+
+
+
+
+
+
 
         <dependency>
             <groupId>org.apache.flink</groupId>
