+++ flink_java/src/main/java/com/dida/item/common/ComonConfig.java
@@ -1,0 +1,17 @@
+package com.dida.item.common;
+/**
+ * @Author：zhd
+ * @Date: 2021/8/22 16:49
+ * @Dscription:
+ */
+
+
+public class ComonConfig {
+    //clickhouse驱动
+    public static final String CLICKHOUSE_DRIVER = "ru.yandex.clickhouse.ClickHouseDriver";
+
+    //clickhouse URL
+    public static final String CLICKHOUSE_URL = "jdbc:clickhouse://10.150.10.15:8123/test";
+
+
+}
