--- flink_java/src/main/java/com/xiaoshouyi/chapter01/UnBoundedStream.java
@@ -1,60 +1,0 @@
-package com.xiaoshouyi.chapter01;
-
-import org.apache.flink.api.common.restartstrategy.RestartStrategies;
-import org.apache.flink.api.common.typeinfo.Types;
-import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.runtime.state.filesystem.FsStateBackend;
-import org.apache.flink.streaming.api.datastream.DataStreamSource;
-import org.apache.flink.streaming.api.datastream.KeyedStream;
-import org.apache.flink.streaming.api.datastream.SingleOutputStreamOperator;
-import org.apache.flink.streaming.api.environment.CheckpointConfig;
-import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.util.Collector;
-
-import java.util.Arrays;
-
-
-public class UnBoundedStream {
-    public static void main(String[] args) throws Exception {
-        // 1. 创建流式执行环境
-        StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
-        //开启checkpoint
-        env.enableCheckpointing(5000);
-        //设置失败重启策略
-        env.setRestartStrategy(RestartStrategies.fixedDelayRestart(3,2000));
-        //设置保证应用异常退出的时候可以保存checkpoint
-        CheckpointConfig checkpointConfig = env.getCheckpointConfig();
-        checkpointConfig.enableExternalizedCheckpoints(CheckpointConfig.ExternalizedCheckpointCleanup.RETAIN_ON_CANCELLATION);
-
-
-
-        //设置checkpoint保存路径
-        env.setStateBackend(new FsStateBackend("s3://state/test/wordCount"));
-//        System.setProperty("s3.endpoint","http://10.150.60.10:31001");
-        System.setProperty("s3.path.style.access","true");
-        System.setProperty("s3.access-key","admin");
-        System.setProperty("s3.secret-key","ZZTnAcs6gf");
-
-
-        // 2. 读取文件
-        DataStreamSource<String> lineDSS = env.socketTextStream("10.150.10.17", 9999);
-        // 3. 转换数据格式
-        SingleOutputStreamOperator<Tuple2<String, Long>> wordAndOne = lineDSS
-                .flatMap((String line, Collector<String> words) -> {
-                    Arrays.stream(line.split(" ")).forEach(words::collect);
-                })
-                .returns(Types.STRING)
-                .map(word -> Tuple2.of(word, 1L))
-                .returns(Types.TUPLE(Types.STRING, Types.LONG));
-        // 4. 分组
-        KeyedStream<Tuple2<String, Long>, String> wordAndOneKS = wordAndOne
-                .keyBy(t -> t.f0);
-        // 5. 求和
-        SingleOutputStreamOperator<Tuple2<String, Long>> result = wordAndOneKS
-                .sum(1);
-        // 6. 打印
-        result.print();
-        // 7. 执行
-        env.execute();
-    }
-}
