--- flink_java/src/main/java/com/xiaoshouyi/practice/pojo/student.java
@@ -1,22 +1,0 @@
-package com.xiaoshouyi.practice.pojo;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.sql.Time;
-
-/**
- * @Author：zhd
- * @Date: 2021/11/4 19:21
- * @Dscription:
- */
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class student {
-    int id;
-    String name;
-    Time date;
-}
