--- flink_java/src/main/java/com/xiaoshouyi/realtime/pojo/UserLogSql.java
@@ -1,18 +1,0 @@
-package com.xiaoshouyi.realtime.pojo;
-
-import com.alibaba.fastjson.JSONObject;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * @Author：zhd
- * @Date: 2021/12/24 13:35
- * @Dscription: 用户行为类
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserLogSql {
- String distinct_Id;
-}
